cre: support native text selection with getTextFromXPointers
This is necessary for the Language Support feature in KOReader.

Before this patch, the only way to get crengine to natively highlight a
string of text was to use getTextFromPositions but this far from ideal
for two reasons:

 1. crengine has some special handling of <ruby> text to try to make it
    harder to select, but the result is that if you try to convert
    XPointers to screen positions and then using getTextFromPositions
    you end up with a selection that doesn't match the XPointers you
    started with. Since the caller already has the XPointers it'd be
    much nicer to not have to go through any conversion at all.

 2. getTextFromPositions cannot create highlights that cross page
    boundaries because the screen position you get from
    getPosFromXpointer is physically not on the screen and thus you end
    up getting nothing from getTextFromPositions.

Being able to optionally have native highlighting from a set of
XPointers would allow the Language Support feature in KOReader much more
smoothly create selections and manage text highlighting when plugins
request the word selection be expanded.

In order to avoid breaking any other callers of getTextFromXPointers, we
default to disabling native selection (after all, most callers probably
aren't interested in selecting the text they're trying to extract).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>