thirdparty: sdcv: update to better support multi-word lookups
Current sdcv doesn't support multi-word lookups in the most ideal way.
In particular, as soon as any word specified in the command-line is
found to not have any dictionary results, it immediately bails. This
makes sdcv quite hard to use for Japanese deinflection and results in
less efficient batching when you have to look up many word candiates.

sdcv has been updated upstream to fix this issue, so update to that
commit.

Ref: https://github.com/Dushistov/sdcv/pull/77
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>